Code cleanup: removed plugin example that used out of date services. added @Deprecated to the http service, as it's been replaced with the httpRpcClient from runtime OS. renamed the surviving plugin example as example plugin. removed @blsemo from CODEOWNERS as he is no longer involved in the project.
