+++ .github/CODEOWNERS
@@ -1,4 +1,4 @@
-* @pmck91 @blsemo
+* @pmck91
 # Build scripts should be audited by BLT
 *.gradle           @corda/blt
 gradle.properties  @corda/corda5-team-leads
