+++ app/src/main/kotlin/net/corda/cli/application/commands/SetCurrentNodeCommand.kt
@@ -1,6 +1,6 @@
 package net.corda.cli.application.commands
 
-import net.corda.cli.application.services.Files
+import net.corda.cli.application.utils.Files
 import org.yaml.snakeyaml.Yaml
 import picocli.CommandLine
 import java.io.FileInputStream
