+++ plugins/example/build.gradle
