+++ plugins/example/gradle.properties
@@ -1,0 +1,7 @@
+version=0.0.1
+
+pluginId=example-plugin
+pluginClass=net.corda.cli.plugins.examples.ExamplePlugin
+pluginProvider=R3
+pluginDescription="An example plugin that shows the use of class based subcommands."
+
