+++ plugins/example/src/main/kotlin/net/corda/cli/plugins/examples/SubCommandOne.kt
@@ -1,0 +1,10 @@
+package net.corda.cli.plugins.examples
+
+import picocli.CommandLine
+
+@CommandLine.Command(name = "sub-command", description = ["Example subcommand."])
+class SubCommandOne : Runnable {
+    override fun run() {
+        println("Hello from the example plugin!")
+    }
+}
