--- plugins/plugin1/build.gradle
@@ -1,19 +1,0 @@
-plugins {
-    id 'org.jetbrains.kotlin.kapt'
-}
-
-dependencies {
-
-    compileOnly project(":api")
-    compileOnly "org.pf4j:pf4j:${pf4jVersion}"
-    kapt "org.pf4j:pf4j:${pf4jVersion}"
-
-    testImplementation project(":api")
-    testImplementation "org.pf4j:pf4j:${pf4jVersion}"
-    testImplementation "org.junit.jupiter:junit-jupiter:5.8.2"
-    testImplementation "com.github.stefanbirkner:system-lambda:1.2.1"
-}
-
-test {
-    useJUnitPlatform()
-}
