--- plugins/plugin1/gradle.properties
@@ -1,6 +1,0 @@
-version=0.0.1
-
-pluginId=example-plugin-one
-pluginClass=net.corda.cli.plugins.examples.ExamplePluginOne
-pluginProvider=R3
-pluginDescription="An example plugin that shows the use of function based subcommands and services loaded from the host."
