--- plugins/plugin2/gradle.properties
@@ -1,7 +1,0 @@
-version=0.0.1
-
-pluginId=example-plugin-two
-pluginClass=net.corda.cli.plugins.examples.ExamplePluginTwo
-pluginProvider=R3
-pluginDescription="An example plugin that shows the use of class based subcommands."
-
