+++ settings.gradle
@@ -116,10 +116,7 @@ rootProject.name = "corda-cli-plugin-host"
 include 'api'
 include 'app'
 include 'plugins'
-include 'plugins:plugin1'
-include 'plugins:plugin2'
-
-
+include 'plugins:example'
 
 gradleEnterprise {
     server = gradleEnterpriseUrl
